# generated/swagger.json
{
  "openapi": "3.0.0",
  "components": {
...
    "schemas": {
...
      "TErrorCause": {
        "$ref": "#/components/schemas/Record_string.unknown_"
      },
      "IHttpErrorDto": {
        "properties": {
          "name": {
            "type": "string"
          },
          "errorKey": {
            "type": "string"
          },
          "validationErrors": {
            "items": {
              "$ref": "#/components/schemas/IApiValidationError"
            },
            "type": "array"
          },
          "cause": {
            "$ref": "#/components/schemas/TErrorCause"
          }
        },
        "required": ["name", "errorKey"],
        "type": "object",
        "additionalProperties": false
      },
...
      "UNormalBalance": {
        "type": "string",
        "enum": ["debit", "credit"]
      },
...
      "ILedgerAccountDto": {
        "properties": {
          "id": {
            "$ref": "#/components/schemas/TEntityId"
          },
          "code": {
            "type": "string"
          },
          "materializedPath": {
            "type": "string"
          },
          "accountingEntityId": {
            "$ref": "#/components/schemas/TEntityId"
          },
          "type": {
            "$ref": "#/components/schemas/ULedgerType"
          },
          "normalBalance": {
            "$ref": "#/components/schemas/UNormalBalance"
          },
          "subType": {},
          "behavior": {
            "$ref": "#/components/schemas/ULedgerAccountBehavior"
          },
          "isControlAccount": {
            "type": "boolean"
          },
          "controlAccountId": {
            "$ref": "#/components/schemas/TEntityId"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ULedgerAccountStatus"
          },
          "contraAccountRule": {
            "$ref": "#/components/schemas/UContraAccountRule"
          },
          "adjunctAccountRule": {
            "$ref": "#/components/schemas/UAdjunctAccountRule"
          },
          "openingBalanceDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdBy": {
            "$ref": "#/components/schemas/TEntityId"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time"
          },
          "balance": {
            "$ref": "#/components/schemas/IMoneyDto"
          },
          "functionalBalance": {
            "$ref": "#/components/schemas/IMoneyDto"
          }
        },
        "required": [
          "id",
          "code",
          "materializedPath",
          "accountingEntityId",
          "type",
          "normalBalance",
          "subType",
          "behavior",
          "isControlAccount",
          "name",
          "status",
          "contraAccountRule",
          "adjunctAccountRule",
          "openingBalanceDate",
          "createdBy",
          "createdAt",
          "updatedAt",
          "balance",
          "functionalBalance"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "IRecommendedBootstrapDto": {
        "properties": {
          "receivables": {
            "items": {
              "$ref": "#/components/schemas/IRecommendedBootstrapAccountDto"
            },
            "type": "array"
          },
          "payables": {
            "items": {
              "$ref": "#/components/schemas/IRecommendedBootstrapAccountDto"
            },
            "type": "array"
          },
          "revenue": {
            "items": {
              "$ref": "#/components/schemas/IRecommendedBootstrapAccountDto"
            },
            "type": "array"
          },
          "expense": {
            "items": {
              "$ref": "#/components/schemas/IRecommendedBootstrapAccountDto"
            },
            "type": "array"
          },
          "suspense": {
            "items": {
              "$ref": "#/components/schemas/IRecommendedBootstrapAccountDto"
            },
            "type": "array"
          }
        },
        "required": [
          "receivables",
          "payables",
          "revenue",
          "expense",
          "suspense"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "IBankAccountCreationReq": {
        "properties": {
          "name": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ULedgerAccountCreationStatus",
            "description": "Defaults to active when omitted."
          },
          "currencyCode": {
            "type": "string"
          },
          "controlAccountId": {
            "type": "string"
          },
          "bankAccount": {
            "$ref": "#/components/schemas/IBankDetailsCreationReq"
          },
          "openingBalance": {
            "allOf": [
              {
                "$ref": "#/components/schemas/IOpeningBalanceDto"
              }
            ],
            "nullable": true
          }
        },
        "required": ["name", "currencyCode", "bankAccount", "openingBalance"],
        "type": "object",
        "additionalProperties": false
      },
...
      "ICreateSuspenseAccountDto": {
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["asset", "liability"]
          },
          "currencyCode": {
            "type": "string"
          }
        },
        "required": ["name", "type", "currencyCode"],
        "type": "object",
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "info": {
    "title": "drimsheet-core",
    "version": "0.1.0-alpha.1",
    "description": "This repo contains the server-side source code for Drimsheet's bookkeeping and tax app.",
    "contact": {
      "name": "Osahon Oboite"
    }
  },
  "paths": {
...
    "/accounts/recommended-bootstrap": {
      "get": {
        "operationId": "getRecommendedBootstrap",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IRecommendedBootstrapDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          }
        },
        "description": "Get recommended posting account setup, grouped with nested sub accounts.",
        "tags": ["Accounts"],
        "security": [],
        "parameters": []
      }
    },
    "/accounts/asset/bank": {
      "post": {
        "operationId": "createBankAccount",
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ILedgerAccountDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          }
        },
        "description": "Create an Active or Draft bank subaccount; status defaults to Active.",
        "tags": ["Asset Accounts", "Accounts"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IBankAccountCreationReq"
              }
            }
          }
        }
      }
    },
    "/accounts/asset/receivables/statutory": {
      "post": {
        "operationId": "createStatutoryReceivableAccount",
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ILedgerAccountDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          }
        },
        "description": "Create an Active or Draft statutory receivable subaccount; status defaults to Active.",
        "tags": ["Asset Accounts", "Accounts"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ICreateStatutoryReceivableAccountDto"
              }
            }
          }
        }
      }
    },
    "/accounts/asset/receivables/trade": {
      "post": {
        "operationId": "createTradeReceivableAccount",
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ILedgerAccountDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          }
        },
        "description": "Create an Active or Draft trade receivable subaccount; status defaults to Active.",
        "tags": ["Asset Accounts", "Accounts"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ICreateTradeReceivableAccountDto"
              }
            }
          }
        }
      }
    },
    "/accounts/expenses": {
      "post": {
        "operationId": "createExpenseAccount",
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ILedgerAccountDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          }
        },
        "description": "Create an Active or Draft expense subaccount; status defaults to Active.",
        "tags": ["Expense Accounts", "Accounts"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ICreateExpenseAccountDto"
              }
            }
          }
        }
      }
    },
    "/accounts/liability/payables/statutory": {
      "post": {
        "operationId": "createStatutoryPayableAccount",
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ILedgerAccountDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          }
        },
        "description": "Create an Active or Draft statutory payable subaccount; status defaults to Active.",
        "tags": ["Liability Accounts", "Accounts"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ICreateStatutoryPayableAccountDto"
              }
            }
          }
        }
      }
    },
    "/accounts/liability/payables/trade": {
      "post": {
        "operationId": "createTradePayableAccount",
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ILedgerAccountDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          }
        },
        "description": "Create an Active or Draft trade payable subaccount; status defaults to Active.",
        "tags": ["Liability Accounts", "Accounts"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ICreateTradePayableAccountDto"
              }
            }
          }
        }
      }
    },
    "/accounts/revenues": {
      "post": {
        "operationId": "createRevenueAccount",
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ILedgerAccountDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          }
        },
        "description": "Create an Active or Draft revenue subaccount; status defaults to Active.",
        "tags": ["Revenue Accounts", "Accounts"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ICreateRevenueAccountDto"
              }
            }
          }
        }
      }
    },
    "/accounts/suspense": {
      "post": {
        "operationId": "createSuspenseAccount",
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ILedgerAccountDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IHttpErrorDto"
                }
              }
            }
          }
        },
        "description": "Create one asset or liability suspense account for the selected currency.",
        "tags": ["Accounts"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ICreateSuspenseAccountDto"
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ]
}